requirements.txt
# Copyright (c) 2025 Sumi Somangili
# All rights reserved.

streamlit>=1.28.0
python-dotenv>=1.0.0
ollama>=0.1.2
langchain-core>=0.1.0
langchain-community>=0.0.20
python-dateutil>=2.8.2
openai>=1.0.0  # For GPT models
langchain-openai>=0.0.8  # OpenAI integration
